Fix the get_stores_by_offset analysis
I'm not sure where we were going wrong before, but we weren't
properly following the use-def or the def-use chain to get all
store operations. The stores for most of the FastIoDispatch
routines were using an alias that we were failing to detect
earlier.

Fix the analysis by first traversing the use-def chain in an
attempt to get the base definition, then following the def-use
chain for all potential aliases in order to get all uses for
this pointer (and therefore all stores using it as a base
address).